@@ -1,63 +1,63 @@
 {
-    "name": "angular-auth-oidc-client",
-    "version": "0.0.0",
-    "scripts": {
-        "ng": "ng",
-        "build": "npm run build-lib",
-        "watch-lib": "ng build angular-auth-oidc-client --watch",
-        "build-lib": "npm run copy-readme-to-lib-folder && ng build angular-auth-oidc-client",
-        "test-lib": "ng test angular-auth-oidc-client",
-        "test-lib-ci": "ng test angular-auth-oidc-client --watch=false --browsers=ChromeHeadless",
-        "lint-lib": "ng lint angular-auth-oidc-client",
-        "pack-lib": "npm run build-lib && npm pack ./dist/angular-auth-oidc-client",
-        "publish-lib": "npm run build-lib && npm publish ./dist/angular-auth-oidc-client",
-        "publish-lib-next": "npm run build-lib && npm publish --tag next ./dist/angular-auth-oidc-client",
-        "compodoc": "compodoc -p tsconfig.json",
-        "compodoc:serve": "compodoc -s",
-        "start": "ng serve --ssl true --project sample-code-flow -o",
-        "copy-readme-to-lib-folder": "copyfiles README.md ./projects/angular-auth-oidc-client"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "~7.2.0",
-        "@angular/common": "~7.2.0",
-        "@angular/compiler": "~7.2.0",
-        "@angular/core": "~7.2.0",
-        "@angular/forms": "~7.2.0",
-        "@angular/platform-browser": "~7.2.0",
-        "@angular/platform-browser-dynamic": "~7.2.0",
-        "@angular/router": "~7.2.0",
-        "core-js": "^2.5.4",
-        "rxjs": "~6.3.3",
-        "tslib": "^1.9.0",
-        "zone.js": "~0.8.26",
-        "jsrsasign": "8.0.12"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "~0.13.0",
-        "@angular-devkit/build-ng-packagr": "~0.13.0",
-        "@angular/cli": "~7.3.1",
-        "@angular/compiler-cli": "~7.2.0",
-        "@angular/language-service": "~7.2.0",
-        "@compodoc/compodoc": "^1.1.9",
-        "@types/jasmine": "~2.8.8",
-        "@types/jasminewd2": "~2.0.3",
-        "@types/node": "^8.9.5",
-        "codelyzer": "~4.5.0",
-        "copyfiles": "^2.1.0",
-        "jasmine-core": "~2.99.1",
-        "jasmine-spec-reporter": "~4.2.1",
-        "karma": "~3.1.1",
-        "karma-chrome-launcher": "~2.2.0",
-        "karma-coverage-istanbul-reporter": "~2.0.1",
-        "karma-jasmine": "~1.1.2",
-        "karma-jasmine-html-reporter": "^0.2.2",
-        "ng-packagr": "^4.2.0",
-        "protractor": "~5.4.0",
-        "ts-node": "~7.0.0",
-        "tsickle": ">=0.34.0",
-        "tslib": "^1.9.0",
-        "tslint": "~5.11.0",
-        "typescript": "~3.2.2"
-    }
+  "name": "angular-auth-oidc-client",
+  "version": "10.0.0",
+  "scripts": {
+    "ng": "ng",
+    "build": "npm run build-lib",
+    "build-app": "ng build sample-code-flow",
+    "watch-lib": "ng build angular-auth-oidc-client --watch",
+    "build-lib": "npm run copy-readme-to-lib-folder && ng build angular-auth-oidc-client",
+    "test-lib": "ng test angular-auth-oidc-client",
+    "test-lib-ci": "ng test angular-auth-oidc-client --watch=false --browsers=ChromeHeadless",
+    "lint-lib": "ng lint angular-auth-oidc-client",
+    "pack-lib": "npm run build-lib && npm pack ./dist/angular-auth-oidc-client",
+    "publish-lib": "npm run build-lib && npm publish ./dist/angular-auth-oidc-client",
+    "publish-lib-next": "npm run build-lib && npm publish --tag next ./dist/angular-auth-oidc-client",
+    "compodoc": "compodoc -p tsconfig.json",
+    "compodoc:serve": "compodoc -s",
+    "start": "ng serve --ssl true --project sample-code-flow -o",
+    "copy-readme-to-lib-folder": "copyfiles README.md ./projects/angular-auth-oidc-client"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "~8.0.0",
+    "@angular/common": "~8.0.0",
+    "@angular/compiler": "~8.0.0",
+    "@angular/core": "~8.0.0",
+    "@angular/forms": "~8.0.0",
+    "@angular/platform-browser": "~8.0.0",
+    "@angular/platform-browser-dynamic": "~8.0.0",
+    "@angular/router": "~8.0.0",
+    "core-js": "^2.5.4",
+    "rxjs": "~6.5.2",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1",
+    "jsrsasign": "8.0.12"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.800.0",
+    "@angular-devkit/build-ng-packagr": "~0.800.0",
+    "@angular/cli": "~8.0.2",
+    "@angular/compiler-cli": "~8.0.0",
+    "@angular/language-service": "~8.0.0",
+    "@compodoc/compodoc": "^1.1.9",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^8.9.5",
+    "codelyzer": "^5.0.1",
+    "copyfiles": "^2.1.0",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~3.1.1",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "ng-packagr": "^5.1.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tsickle": "^0.35.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.4.5"
+  }
 }
