@@ -1,15 +1,15 @@
 // Public classes.
 
+export * from './models/auth.configuration';
 export * from './models/auth.well-known-endpoints';
 export * from './models/authorization-result';
 export * from './models/authorization-state.enum';
 export * from './models/jwtkeys';
 export * from './models/validate-state-result.model';
 export * from './models/validation-result.enum';
-export * from './modules/auth.configuration';
 export * from './modules/auth.module';
+export * from './services/oidc-token-helper.service';
 export * from './services/oidc.security.config.service';
 export * from './services/oidc.security.service';
 export * from './services/oidc.security.storage';
 export * from './services/oidc.security.validation';
-export * from './services/oidc-token-helper.service';
