@@ -1,22 +1,22 @@
 import { Injectable } from '@angular/core';
-import { AuthConfiguration } from '../modules/auth.configuration';
+import { ConfigurationProvider } from './auth-configuration.provider';
 
 @Injectable()
 export class LoggerService {
-    constructor(private authConfiguration: AuthConfiguration) {}
+    constructor(private configurationProvider: ConfigurationProvider) {}
 
     logError(message: any) {
         console.error(message);
     }
 
     logWarning(message: any) {
-        if (this.authConfiguration.isLogLevelWarningEnabled) {
+        if (this.configurationProvider.openIDConfiguration.log_console_warning_active) {
             console.warn(message);
         }
     }
 
     logDebug(message: any) {
-        if (this.authConfiguration.isLogLevelDebugEnabled) {
+        if (this.configurationProvider.openIDConfiguration.log_console_debug_active) {
             console.log(message);
         }
     }
