@@ -2,8 +2,9 @@ import { HttpClientModule } from '@angular/common/http';
 import { TestBed } from '@angular/core/testing';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouterTestingModule } from '@angular/router/testing';
-import { AuthConfiguration, OpenIDImplicitFlowConfiguration } from '../../lib/modules/auth.configuration';
+import { OpenIdConfiguration } from '../../lib/models/auth.configuration';
 import { AuthModule } from '../../lib/modules/auth.module';
+import { ConfigurationProvider } from '../../lib/services/auth-configuration.provider';
 import { EqualityHelperService } from '../../lib/services/oidc-equality-helper.service';
 import { LoggerService } from '../../lib/services/oidc.logger.service';
 import { OidcSecurityStorage } from '../../lib/services/oidc.security.storage';
@@ -13,13 +14,13 @@ import { TestStorage } from '../common/test-storage.service';
 
 describe('OidcSecurityValidation', () => {
     let oidcSecurityValidation: OidcSecurityValidation;
-    let authConfiguration: AuthConfiguration;
+    let configProvider: ConfigurationProvider;
 
     beforeEach(() => {
         TestBed.configureTestingModule({
             imports: [BrowserModule, HttpClientModule, RouterTestingModule, AuthModule.forRoot()],
             providers: [
-                AuthConfiguration,
+                ConfigurationProvider,
                 EqualityHelperService,
                 OidcSecurityValidation,
                 {
@@ -36,64 +37,64 @@ describe('OidcSecurityValidation', () => {
 
     beforeEach(() => {
         oidcSecurityValidation = TestBed.get(OidcSecurityValidation);
-        authConfiguration = TestBed.get(AuthConfiguration);
+        configProvider = TestBed.get(ConfigurationProvider);
     });
 
     it('validate aud string', () => {
-        let openIDImplicitFlowConfiguration = new OpenIDImplicitFlowConfiguration();
-        openIDImplicitFlowConfiguration.stsServer = 'https://localhost:5001';
-        openIDImplicitFlowConfiguration.redirect_url = 'https://localhost:44386';
-        openIDImplicitFlowConfiguration.client_id = '188968487735-b1hh7k87nkkh6vv84548sinju2kpr7gn.apps.googleusercontent.com';
-        openIDImplicitFlowConfiguration.response_type = 'id_token token';
-        openIDImplicitFlowConfiguration.scope = 'openid email profile';
-        openIDImplicitFlowConfiguration.post_logout_redirect_uri = 'https://localhost:44386/Unauthorized';
-        openIDImplicitFlowConfiguration.post_login_route = '/home';
-        openIDImplicitFlowConfiguration.forbidden_route = '/Forbidden';
-        openIDImplicitFlowConfiguration.unauthorized_route = '/Unauthorized';
-        openIDImplicitFlowConfiguration.start_checksession = false;
-        openIDImplicitFlowConfiguration.silent_renew = false;
-        openIDImplicitFlowConfiguration.silent_renew_offset_in_seconds = 0;
-        openIDImplicitFlowConfiguration.log_console_warning_active = true;
-        openIDImplicitFlowConfiguration.log_console_debug_active = true;
-        openIDImplicitFlowConfiguration.max_id_token_iat_offset_allowed_in_seconds = 10;
+        const config: OpenIdConfiguration = {};
+        config.stsServer = 'https://localhost:5001';
+        config.redirect_url = 'https://localhost:44386';
+        config.client_id = '188968487735-b1hh7k87nkkh6vv84548sinju2kpr7gn.apps.googleusercontent.com';
+        config.response_type = 'id_token token';
+        config.scope = 'openid email profile';
+        config.post_logout_redirect_uri = 'https://localhost:44386/Unauthorized';
+        config.post_login_route = '/home';
+        config.forbidden_route = '/Forbidden';
+        config.unauthorized_route = '/Unauthorized';
+        config.start_checksession = false;
+        config.silent_renew = false;
+        config.silent_renew_offset_in_seconds = 0;
+        config.log_console_warning_active = true;
+        config.log_console_debug_active = true;
+        config.max_id_token_iat_offset_allowed_in_seconds = 10;
 
-        authConfiguration.init(openIDImplicitFlowConfiguration);
+        configProvider.setup(config, null);
 
         const dataIdToken = { aud: 'banana' };
-        let valueTrue = oidcSecurityValidation.validate_id_token_aud(dataIdToken, 'banana');
+        const valueTrue = oidcSecurityValidation.validate_id_token_aud(dataIdToken, 'banana');
         expect(valueTrue).toEqual(true);
 
-        let valueFalse = oidcSecurityValidation.validate_id_token_aud(dataIdToken, 'bananammmm');
+        const valueFalse = oidcSecurityValidation.validate_id_token_aud(dataIdToken, 'bananammmm');
         expect(valueFalse).toEqual(false);
     });
 
     it('validate aud array', () => {
-        let openIDImplicitFlowConfiguration = new OpenIDImplicitFlowConfiguration();
-        openIDImplicitFlowConfiguration.stsServer = 'https://localhost:5001';
-        openIDImplicitFlowConfiguration.redirect_url = 'https://localhost:44386';
-        openIDImplicitFlowConfiguration.client_id = '188968487735-b1hh7k87nkkh6vv84548sinju2kpr7gn.apps.googleusercontent.com';
-        openIDImplicitFlowConfiguration.response_type = 'id_token token';
-        openIDImplicitFlowConfiguration.scope = 'openid email profile';
-        openIDImplicitFlowConfiguration.post_logout_redirect_uri = 'https://localhost:44386/Unauthorized';
-        openIDImplicitFlowConfiguration.post_login_route = '/home';
-        openIDImplicitFlowConfiguration.forbidden_route = '/Forbidden';
-        openIDImplicitFlowConfiguration.unauthorized_route = '/Unauthorized';
-        openIDImplicitFlowConfiguration.start_checksession = false;
-        openIDImplicitFlowConfiguration.silent_renew = false;
-        openIDImplicitFlowConfiguration.silent_renew_offset_in_seconds = 0;
-        openIDImplicitFlowConfiguration.log_console_warning_active = true;
-        openIDImplicitFlowConfiguration.log_console_debug_active = true;
-        openIDImplicitFlowConfiguration.max_id_token_iat_offset_allowed_in_seconds = 10;
+        const config: OpenIdConfiguration = {};
+        config.stsServer = 'https://localhost:5001';
+        config.redirect_url = 'https://localhost:44386';
+        config.client_id = '188968487735-b1hh7k87nkkh6vv84548sinju2kpr7gn.apps.googleusercontent.com';
+        config.response_type = 'id_token token';
+        config.scope = 'openid email profile';
+        config.post_logout_redirect_uri = 'https://localhost:44386/Unauthorized';
+        config.post_login_route = '/home';
+        config.forbidden_route = '/Forbidden';
+        config.unauthorized_route = '/Unauthorized';
+        config.start_checksession = false;
+        config.silent_renew = false;
+        config.silent_renew_offset_in_seconds = 0;
+        config.log_console_warning_active = true;
+        config.log_console_debug_active = true;
+        config.max_id_token_iat_offset_allowed_in_seconds = 10;
 
-        authConfiguration.init(openIDImplicitFlowConfiguration);
+        configProvider.setup(config, null);
 
         const dataIdToken = {
             aud: ['banana', 'apple', 'https://nice.dom'],
         };
-        let valueTrue = oidcSecurityValidation.validate_id_token_aud(dataIdToken, ['banana', 'apple', 'https://nice.dom']);
+        const valueTrue = oidcSecurityValidation.validate_id_token_aud(dataIdToken, ['banana', 'apple', 'https://nice.dom']);
         expect(valueTrue).toEqual(true);
 
-        let valueFalse = oidcSecurityValidation.validate_id_token_aud(dataIdToken, ['ooo', 'apple', 'https://nice.dom']);
+        const valueFalse = oidcSecurityValidation.validate_id_token_aud(dataIdToken, ['ooo', 'apple', 'https://nice.dom']);
         expect(valueFalse).toEqual(false);
     });
 });
