@@ -2,13 +2,7 @@ import { HttpClientModule } from '@angular/common/http';
 import { APP_INITIALIZER, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
-import {
-    AuthModule,
-    AuthWellKnownEndpoints,
-    OidcConfigService,
-    OidcSecurityService,
-    OpenIDImplicitFlowConfiguration,
-} from 'angular-auth-oidc-client';
+import { AuthModule, ConfigResult, OidcConfigService, OidcSecurityService, OpenIdConfiguration } from 'angular-auth-oidc-client';
 import { AppComponent } from './app.component';
 
 export function loadConfig(oidcConfigService: OidcConfigService) {
@@ -41,17 +35,17 @@ export function loadConfig(oidcConfigService: OidcConfigService) {
 })
 export class AppModule {
     constructor(private oidcSecurityService: OidcSecurityService, private oidcConfigService: OidcConfigService) {
-        this.oidcConfigService.onConfigurationLoaded.subscribe(() => {
-            const config = new OpenIDImplicitFlowConfiguration();
-            config.stsServer = 'https://offeringsolutions-sts.azurewebsites.net';
-            config.redirect_url = 'https://localhost:4200';
-            config.client_id = 'angularClient';
-            config.scope = 'openid profile email';
-            config.response_type = 'code';
-
-            config.silent_renew = true;
-            config.silent_renew_url = 'https://localhost:4200/silent-renew.html';
-            config.log_console_debug_active = true;
+        this.oidcConfigService.onConfigurationLoaded.subscribe((configResult: ConfigResult) => {
+            const config: OpenIdConfiguration = {
+                stsServer: configResult.customConfig.stsServer,
+                redirect_url: 'https://localhost:4200',
+                client_id: 'angularClient',
+                scope: 'openid profile email',
+                response_type: 'code',
+                silent_renew: true,
+                silent_renew_url: 'https://localhost:4200/silent-renew.html',
+                log_console_debug_active: true,
+            };
 
             //config.start_checksession = true;
             //config.post_login_route = '/home';
@@ -60,10 +54,7 @@ export class AppModule {
             //config.max_id_token_iat_offset_allowed_in_seconds = 5;
             //config.history_cleanup_off = true;
 
-            const authWellKnownEndpoints = new AuthWellKnownEndpoints();
-            authWellKnownEndpoints.setWellKnownEndpoints(this.oidcConfigService.wellKnownEndpoints);
-
-            this.oidcSecurityService.setupModule(config, authWellKnownEndpoints);
+            this.oidcSecurityService.setupModule(config, configResult.authWellknownEndpoints);
         });
     }
 }
